requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-dotenv
pytest